requirements.txt
streamlit==1.43.0
gitpython==3.1.41
docker==6.1.3
python-dotenv==1.0.1
pydantic==2.6.1
humanize==4.9.0
websocket-client>=0.32.0
streamlit-ace==0.1.1
